KokkosBatched - cuda perf test of dense blocks on white (p100)
